MemExceed not solved

+++ Assignment_4/src/main.cpp
@@ -178,7 +178,7 @@ struct centroid {
     double x;
     double y;
     double z;
-    long idx; // index
+    int idx; // index
 } centroid;
 
 // sorting criteria.
